Add another test for test_get_student_unions and move to separate file
